+++ app/__init__.py
@@ -9,4 +9,4 @@ content.register_model('page', model.Page, 'pages')
 
 # Application settings form
 permission.define_permission('app.settings.manage', 'manage_app_settings', 'app')
-settings.define('app', settings_form.Form, __name__ + '@application', 'fa fa-cubes', 'app.settings.manage', -999)
+settings.define('app', settings_form.Form, __name__ + '@application', 'fa fa-cube', 'app.settings.manage', -999)