+++ app/settings_form.py
@@ -1,7 +1,7 @@
 """
 """
-from datetime import datetime as _datetime
-from pytsite import widget as _widget, lang as _lang, validation as _validation, settings
+from datetime import datetime
+from pytsite import widget, lang, validation, settings
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -10,20 +10,20 @@ __license__ = 'MIT'
 
 class Form(settings.Form):
     def _setup_widgets(self):
-        self.add_widget(_widget.input.StringList(
+        self.add_widget(widget.input.StringList(
             uid='setting_links',
             weight=10,
-            label=_lang.t('links'),
-            add_btn_label=_lang.t('add_link'),
+            label=lang.t('links'),
+            add_btn_label=lang.t('add_link'),
             unique=True,
-            rules=_validation.rule.Url(),
+            rules=validation.rule.Url(),
         ))
 
-        current_year = _datetime.now().year
-        self.add_widget(_widget.input.Integer(
+        current_year = datetime.now().year
+        self.add_widget(widget.input.Integer(
             uid='setting_launch_year',
             weight=20,
-            label=_lang.t('launch_year'),
+            label=lang.t('launch_year'),
             h_size='col-xs-12 col-sm-3 col-md-2 col-lg-1',
             required=True,
             min=1984,